Update eth-migrate-archiveLocation.js
As far as I can see, the script has the problem that all archive locations will be moved to extras. However, very occasionally archive locations are used for other things. This isn't the case in the main libraries we use, however, there are many many libraries I have and some of them may have data in the archive location. I've added some pseudo-code above.

Also, it may be the case that the script has to run more than once. I've added some pseudo-code to allow that.